+++ src/index.py
@@ -18,7 +18,7 @@ from starlette.exceptions import HTTPException as StarletteHTTPException
 import uvicorn
 
 from .infrastructure.translation_service import translation_service
-from infrastructure.i18n import I18nConfig
+from src.infrastructure.i18n import I18nConfig
 from src.infrastructure.config import settings
 from src.infrastructure.web.home_controller import router as home_router
 from src.infrastructure.web.status_controller import router as status_router
